Split database model in multiple files
