+++ src/models/__init__.py
@@ -1,0 +1,11 @@
+from .database import db, connect_db
+from .user import User
+from .book import Book
+from .user_book import UserBook
+from .comment import Comment
+from .club import Club
+from .club_book import ClubBook
+from .club_member import ClubMembers
+from .message import Message
+
+__all__=["User", "Book", "UserBook", "Comment", "Club", "ClubBook", "ClubMembers", "Message", "db", "connect_db"]
