+++ src/models/database.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+db = SQLAlchemy()
+
+def connect_db(app):
+    """Function to connect flask app to the database."""
+    db.app = app
+    db.init_app(app)
